A Bitset Class

This is a data strucutre that compactly stores the presence of integers in some finite set (max_size), and can efficiently perform set operations (union, intersection, complement). WARNING: all operations (except $not) are in-place so please use $copy if you would like to perform an operation without destroying your current bitset.

Public fields

.bitset

a pointer to the underlying IterableBitset

max_size

the maximum size of the bitset

Methods

Public methods

Bitset$new()

Bitset$insert()

Bitset$remove()

Bitset$size()

Bitset$or()

Bitset$and()

Bitset$not()

Bitset$xor()

Bitset$set_difference()

Bitset$sample()

Bitset$copy()

Bitset$to_vector()

Bitset$clone()

Method new()

create a bitset

Usage

Bitset$new(size, from = NULL)

Arguments

size

the size of the bitset

from

pointer to an existing IterableBitset to use

Method insert()

insert into the bitset

Usage

Bitset$insert(v)

Arguments

v

an integer vector of elements to insert

Method remove()

remove from the bitset

Usage

Bitset$remove(v)

Arguments

v

an integer vector of elements (not indices) to remove

Method size()

the number of elements in the set

Usage

Bitset$size()

Method or()

to "bitwise or" or union two Bitsets

Usage

Bitset$or(other)

Arguments

other

the other bitset to combine

Method and()

to "bitwise and" or intersect two Bitsets

Usage

Bitset$and(other)

Arguments

other

the other bitset to combine

Method not()

to "bitwise not" or complement a Bitset This method returns a new Bitset rather than doing in-place modification.

Usage

Bitset$not()

Method xor()

to "bitwise xor" or get the symmetric difference of two Bitsets (keep elements in either Bitset but not in their intersection)

Usage

Bitset$xor(other)

Method set_difference()

Take the set difference of this Bitset with another (keep elements of this Bitset which are not in other).

Usage

Bitset$set_difference(other)

Method sample()

to sample a Bitset

Usage

Bitset$sample(rate)

Arguments

rate

the success probability for keeping each element, can be a single value for all elements or a vector with of unique probabilities for keeping each element

Method copy()

returns a copy the bitset

Usage

Bitset$copy()

Method to_vector()

return an integer vector representing the elements which are set

Usage

Bitset$to_vector()

Method clone()

The objects of this class are cloneable with this method.

Usage

Bitset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.