DoubleVariable Class

Represents a continuous variable for an individual in our simulation

Methods

Public methods

DoubleVariable$new()

DoubleVariable$get_values()

DoubleVariable$get_index_of()

DoubleVariable$get_size_of()

DoubleVariable$queue_update()

DoubleVariable$.update()

DoubleVariable$clone()

Method new()

Create a new DoubleVariable

Usage

DoubleVariable$new(initial_values)

Arguments

initial_values

a numeric vector of the initial value for each individual

Method get_values()

get the variable values

Usage

DoubleVariable$get_values(index = NULL)

Arguments

index

optionally return a subset of the variable vector. If NULL, return all values; if passed a Bitset or integer vector, return values of those individuals.

Method get_index_of()

return a Bitset giving individuals whose value lies in an interval Search for indices corresponding to values in the interval [a,b].

Usage

DoubleVariable$get_index_of(a, b)

Arguments

a

lower bound

b

upper bound

Method get_size_of()

return the number of individuals whose value lies in an interval Count individuals whose value lies in the interval [a,b].

Usage

DoubleVariable$get_size_of(a, b)

Arguments

a

lower bound

b

upper bound

Method queue_update()

Queue an update for a variable. There are 4 types of variable update:

1. Subset update. The index vector represents a subset of the variable to update. The value vector, of the same size, represents the new values for that subset 2. Subset fill. The index vector represents a subset of the variable to update. The value vector, of size 1, will fill the specified subset 3. Variable reset. The index vector is set to NULL and the value vector replaces all of the current values in the simulation. The value vector is should match the size of the population. 4. Variable fill. The index vector is set to NULL and the value vector, of size 1, is used to fill all of the variable values in the population.

Usage

DoubleVariable$queue_update(values, index = NULL)

Arguments

values

a vector or scalar of values to assign at the index

index

is the index at which to apply the change, use NULL for the fill options. If using indices, this may be either a vector of integers or a Bitset.

Method .update()

Usage

DoubleVariable$.update()

Method clone()

The objects of this class are cloneable with this method.

Usage

DoubleVariable$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.