Event Class

Describes a general event in the simulation

Methods

Public methods

Event$new()

Event$add_listener()

Event$schedule()

Event$clear_schedule()

Event$.tick()

Event$.process()

Event$.process_listener()

Event$.process_listener_cpp()

Event$clone()

Method new()

Initialise an Event

Usage

Event$new()

Arguments

name,

the name of the event

Method add_listener()

Add an event listener

Usage

Event$add_listener(listener)

Arguments

listener

the function to be executed on the event

Method schedule()

schedule this event to occur in the future

Usage

Event$schedule(delay)

Arguments

delay

the number of timesteps to wait before triggering the event, can be a scalar or an vector of values for several times in the future

Method clear_schedule()

Stop a future event from triggering

Usage

Event$clear_schedule()

Method .tick()

Usage

Event$.tick()

Method .process()

Usage

Event$.process()

Method .process_listener()

Usage

Event$.process_listener(listener)

Method .process_listener_cpp()

Usage

Event$.process_listener_cpp(listener)

Method clone()

The objects of this class are cloneable with this method.

Usage

Event$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.