EventBase Class

Common functionality shared between simple and targeted events.

Methods

Public methods

EventBase$add_listener()

EventBase$.timestep()

EventBase$.tick()

EventBase$.process()

EventBase$clone()

Method add_listener()

Add an event listener.

Usage

EventBase$add_listener(listener)

Arguments

listener

the function to be executed on the event, which takes a single argument giving the time step when this event is triggered.

Method .timestep()

Usage

EventBase$.timestep()

Method .tick()

Usage

EventBase$.tick()

Method .process()

Usage

EventBase$.process()

Method clone()

The objects of this class are cloneable with this method.

Usage

EventBase$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.