Render

Class to render output for the simulation

Methods

Public methods

Render$new()

Render$set_default()

Render$render()

Render$to_dataframe()

Render$clone()

Method new()

Initialise a renderer for the simulation, creates the default state renderers

Usage

Render$new(timesteps)

Arguments

timesteps

number of timesteps in the simulation

Method set_default()

Set a default value for a rendered output renderers

Usage

Render$set_default(name, value)

Arguments

name

the variable to set a default for

value

the default value to set for a variable

Method render()

Update the render with new simulation data

Usage

Render$render(name, value, timestep)

Arguments

name

the variable to render

value

the value to store for the variable

timestep

the timestep of the data point

Method to_dataframe()

Return the render as a data.frame

Usage

Render$to_dataframe()

Method clone()

The objects of this class are cloneable with this method.

Usage

Render$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.