Class: SimAPI The entry point for models to inspect and manipulate the simulation

Class: SimAPI The entry point for models to inspect and manipulate the simulation

Class: SimAPI The entry point for models to inspect and manipulate the simulation

Methods

Public methods

SimAPI$get_state()

SimAPI$get_variable()

SimAPI$queue_state_update()

SimAPI$queue_variable_update()

SimAPI$schedule()

SimAPI$get_scheduled()

SimAPI$clear_schedule()

SimAPI$get_timestep()

SimAPI$get_parameters()

SimAPI$render()

SimAPI$new()

SimAPI$clone()

Method get_state()

Get the index of individuals with a particular state

Usage

SimAPI$get_state(individual, ...)

Arguments

individual

of interest

...

the states of interest

Method get_variable()

Get a variable vector for an individual

Usage

SimAPI$get_variable(individual, variable, index = NULL)

Arguments

individual

the individual of interest

variable

the variable of interest

index

optionally return a subset of the variable vector

Method queue_state_update()

Queue a state update for the end of the timestep

Usage

SimAPI$queue_state_update(individual, state, index)

Arguments

individual

the individual of interest

state

the target state

index

the index of individuals to move to the target state

Method queue_variable_update()

Queue an update for a variable. There are 4 types of variable update:

1. Subset update. The index vector represents a subset of the variable to update. The value vector, of the same size, represents the new values for that subset 2. Subset fill. The index vector represents a subset of the variable to update. The value vector, of size 1, will fill the specified subset 3. Variable reset. The index vector is set to `NULL` and the value vector replaces all of the current values in the simulation. The value vector is should match the size of the population. 4. Variable fill. The index vector is set to `NULL` and the value vector, of size 1, is used to fill all of the variable values in the population.

Usage

SimAPI$queue_variable_update(individual, variable, values, index = numeric(0))

Arguments

individual

is the type of individual to update

variable

a Variable object representing the variable to change

values

a vector or scalar of values to assign at the index

index

is the index at which to apply the change, use NULL for the fill options

Method schedule()

Schedule an event to occur in the future

Usage

SimAPI$schedule(event, target, delay)

Arguments

event

the event to schedule

target

the individuals to pass to the listener

delay

the number of timesteps to wait before triggering the event

Method get_scheduled()

Get the individuals who are scheduled for a particular event

Usage

SimAPI$get_scheduled(event)

Arguments

event,

the event of interest

Method clear_schedule()

Stop a future event from triggering for a subset of individuals

Usage

SimAPI$clear_schedule(event, target)

Arguments

event

the event to stop

target

the individuals to clear

Method get_timestep()

Get the current timestep of the simulation

Usage

SimAPI$get_timestep()

Method get_parameters()

Get the parameters of the simulation

Usage

SimAPI$get_parameters()

Method render()

Add a value for the simulation output

Usage

SimAPI$render(name, value, timestep = NULL)

Arguments

name

the column name

value

the value to assign

timestep

optionally add the value to a specific timestep

Method new()

Create an R wrapper for the API

Usage

SimAPI$new(cpp_api, parameters, renderer)

Arguments

cpp_api

the cpp implementation of the simulation api

parameters

model parameters

renderer

renderer to store model outputs to

Method clone()

The objects of this class are cloneable with this method.

Usage

SimAPI$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.