Event Class

Describes a general event in the simulation.

Super class

individual::EventBase -> Event

Methods

Public methods

Event$new()

Event$schedule()

Event$clear_schedule()

Event$.process_listener()

Event$.process_listener_cpp()

Event$.resize()

Event$.checkpoint()

Event$.restore()

Event$clone()

Inherited methods

individual::EventBase$.process()

individual::EventBase$.tick()

individual::EventBase$.timestep()

individual::EventBase$add_listener()

Method new()

Initialise an Event.

Usage

Event$new()

Method schedule()

Schedule this event to occur in the future.

Usage

Event$schedule(delay)

Arguments

delay

the number of time steps to wait before triggering the event, can be a scalar or a vector of values for events that should be triggered multiple times.

Method clear_schedule()

Stop a future event from triggering.

Usage

Event$clear_schedule()

Method .process_listener()

Usage

Event$.process_listener(listener)

Method .process_listener_cpp()

Usage

Event$.process_listener_cpp(listener)

Method .resize()

Usage

Event$.resize()

Method .checkpoint()

Usage

Event$.checkpoint()

Method .restore()

Usage

Event$.restore(time, schedule)

Method clone()

The objects of this class are cloneable with this method.

Usage

Event$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.